{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = ""
authors = ["{{ cookiecutter.email }}"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.5"
langserve = "^0.2.2"
openai = "^1.34.0"
pydantic = "^2.7.4"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
sse-starlette = "^2.1.2"
pytest-cov = "^5.0.0"
langchain-qdrant = "^0.1.1"
celery = "^5.4.0"
redis = "^5.0.6"
SQLAlchemy = "^2.0.31"
flower = "^2.0.1"
scikit-learn = "^1.5.0"
loguru = "^0.7.2"
gunicorn = "^22.0.0"
langfuse = "^2.36.1"
langchain-openai = "^0.1.8"
psycopg2 = "^2.9.9"
pydantic-settings = "^2.3.3"
langchain-community = "^0.2.5"
pandas = "^2.2.2"
jwcrypto = "^1.5.6"
python-jose = "^3.3.0"
httplib2 = "^0.22.0"
alembic = "^1.13.1"
uvicorn = "^0.24.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
    # omit anything in a .local directory anywhere
    "*/.local/*",
    # omit everything in /usr
    "/usr/*",
    "*/libs/*",
]

[tool.pytest.ini_options]
addopts = "-v --cov-report term --cov-report html:htmlcov --cov-report xml --cov-fail-under=80 --cov ./"
